moved the physics files to their own file and started a rewrite of the rigid body class
upon thinking about it I've decided that the rigid body class doesn't own the sprite that is displayed by it. the sprite should be owned by a child GameObject class. That is the class that you should inherit from. I was also adding support for my vector2D class witch cased the code to be hard to read and for there to be a lot of repeated code. also moved the enums that I had to their own file.. this allows the code to be cleaner and allows for better readability.