changed defintion of tick member function and added non-dynamic objet support
I feel  ScreenH and ScreenW better represent what the numbers mean this will be better in the long run as it helps with readability for people who are new to using my game engine.
I also wrapped the default tick in a if statment checking if the object is dynamic or not if it is then it allows it to move otherwise tick does nothing. The objects nature should be handled by the Window object in the updateAll Function but this also works