Create Error.hpp
this adds the template struct type Error_or_t this is used for returning either an error or the given type.
if the member error is not Error_t::OK then IsDataValid returns false.
if IsDataValid returns false the value returned by GetVal is undefined.
SKC_perror prints a errormsg based off of the Error_t in the passed in Error_orit also takes an optional lineNumber if