reimplemented Console using vfprintf
vfprintf is a form of fprintf that instead of being variadic it takes in a va_list. this allows you to make your functions "printf like" where that means that it takes in a format string and any number of vars then outputs them in the same way as printf would. This greatly simplified this API