Create Console.hpp
this API supersedes skele_Console_lib which is now depreciated...
this is a cross platform API for printing to the screen.
This API allows you to color the screen using ANSI escape codes (in both 8 and 24 bit modes [where supported] ). and comes  with some Functions that format before they print...
this API Also allows you to clear the screen and move the cursor.
This API was tested and is working in Windows terminal...
according to this page https://manpages.ubuntu.com/manpages/trusty/en/man4/console_codes.4.html
 this should work in ubuntu as well...